refactor(frontend): extract pagination into dedicated component
Extract pagination logic from VoskVocabularyTable into a reusable Pagination
component for better separation of concerns and reusability.
